Create ad-search.sh
In this script, the ldapsearch command is used to query the active directory hosts that were found in the previous step. The output of this command is also logged in the file ad_scan.log along with the other messages from the script. This script also checks for the success or failure of the scan and whether any active directory hosts were found. You can modify the script to suit your needs. For example, you can change the target network, the output file, the log file, and the search query as needed.